targeting-service/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.28.1